+++ Modules/Accordion/resources/views/templates/skin-3.blade.php
@@ -44,15 +44,15 @@
         @endphp
 
         <div class="card">
-            <div class="card-header" id="header-item-{{ $edit_field_key }}">
+            <div class="card-header" id="header-item-{{ $params['id'] }}-{{ $edit_field_key }}">
                 <h5 class="mb-0">
-                    <button class="btn btn-link" data-bs-toggle="collapse" data-bs-target="#collapse-accordion-item-{{ $edit_field_key . '-' . $key }}" aria-expanded="true" aria-controls="collapse-accordion-item-{{ $edit_field_key . '-' . $key }}">
+                    <button class="btn btn-link" data-bs-toggle="collapse" data-bs-target="#collapse-accordion-item-{{ $params['id'] }}-{{ $edit_field_key }}-{{ $key }}" aria-expanded="true" aria-controls="collapse-accordion-item-{{ $params['id'] }}-{{ $edit_field_key }}-{{ $key }}">
                         {!! isset($slide['icon']) ? icon_html($slide['icon']) . ' ' : '' !!} <span class="mw-accordion-text-color"> {{ isset($slide['title']) ? $slide['title'] : '' }}</span>
                     </button>
                 </h5>
             </div>
 
-            <div id="collapse-accordion-item-{{ $edit_field_key . '-' . $key }}" class="collapse {{ $key == 0 ? 'show' : '' }}" aria-labelledby="header-item-{{ $edit_field_key }}" data-bs-parent="#mw-accordion-module-{{ $params['id'] }}">
+            <div id="collapse-accordion-item-{{ $params['id'] }}-{{ $edit_field_key }}-{{ $key }}" class="collapse {{ $key == 0 ? 'show' : '' }}" aria-labelledby="header-item-{{ $params['id'] }}-{{ $edit_field_key }}" data-bs-parent="#mw-accordion-module-{{ $params['id'] }}">
                 <div class="card-body mw-accordion-module-content">
                     @include('modules.accordion::partials.render_accordion_item_content')